Code refactor

+++ FEHDataExtractor/GenericExtractionUtilities.cs
@@ -171,7 +171,7 @@ public class ExtractUtils
 
     public static long getLong(long a, byte[] data)
     {
-        return (long)data[a] + (data[a + 1] << 8) + (data[a + 2] << 16) + (data[a + 3] << 24) + ((long)data[a + 4] << 32) + ((long)data[a + 5] << 40) + ((long)data[a + 6] << 48) + ((long)data[a + 7] << 56);
+        return (long)data[a] + ((long)data[a + 1] << 8) + ((long)data[a + 2] << 16) + ((long)data[a + 3] << 24) + ((long)data[a + 4] << 32) + ((long)data[a + 5] << 40) + ((long)data[a + 6] << 48) + ((long)data[a + 7] << 56);
     }
 
     public static String BitmaskConvertToString(UInt32 value, String[] Names) {
